TypeSignatureBuilder tests must use unrestricted classloader
